feat: add vertical vim movement support
This includes j to scroll down and k to scroll up, also supports Ctrl-D
and Ctrl-U for even faster scroll, and support gg (scroll to top) and G
(scroll to bottom)

Yes, by supporting Ctrl-D and Ctrl-U, will prevent the browser defaults.
Ctrl-U by default view the page source (tested on chromium & firefox)
and Ctrl-D by default bookmark the page.

I want this to be disabled by default, and can be enabled by the user
(will be notified if the user hasn't set one)  using the `:set` command,
and store it into `localStorage`. So a TODO for me ig

This was a lot of fun